+++ src/WtBtCore/HftMocker.cpp
@@ -341,7 +341,7 @@ void HftMocker::procOrder(uint32_t localid)
 		return;
 
 	double curPx = curTick->price();
-	uint32_t orderQty = ordInfo._isBuy ? curTick->askqty(0) : curTick->bidqty(0);	//看对手盘的数量
+	double orderQty = ordInfo._isBuy ? curTick->askqty(0) : curTick->bidqty(0);	//看对手盘的数量
 	if (!_use_newpx)
 	{
 		curPx = ordInfo._isBuy ? curTick->askprice(0) : curTick->bidprice(0);
